+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
--   implement v2 Transactions ([#27])
+-   added v2 Transaction support ([#27])
+-   added VendorField display support ([#29])
 
 ### Changed
 -   merged updates from LedgerHQ/ledger-app-ark ([#23])
@@ -17,4 +18,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [#23]: https://github.com/ArkEcosystem/ledger/pull/23
 [#25]: https://github.com/ArkEcosystem/ledger/pull/25
 [#27]: https://github.com/ArkEcosystem/ledger/pull/27
+[#29]: https://github.com/ArkEcosystem/ledger/pull/29
 [unreleased]: https://github.com/ArkEcosystem/ledger/compare/master...develop